[refactor] add social login util with context
